+++ .eslintrc.js
@@ -2,8 +2,11 @@ module.exports = {
     root: true,
     parser: '@typescript-eslint/parser',
     parserOptions: {
-        tsconfigRootDir: __dirname,
-        project: ['./tsconfig.common.json'],
+        // Specify a tsconfig specifically for eslint with simple glob patterns without project references.
+        // Without this checking individual files with lint-staged fails.
+        // https://github.com/typescript-eslint/typescript-eslint/issues/967#issuecomment-531817014
+        // TODO: Debug why extending tsconfig.common.json in the following file doesn't work
+        project: './tsconfig.eslint.json',
     },
     plugins: [
         '@typescript-eslint',