+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# (2020-04-07)
+
+### Bug Fixes
+
+-   **assert:** fix stale import of a11y config ([94c9815](https://github.com/salesforce/sa11y/commit/94c9815d947325c73f312af69c18939e92fd6878))
+-   **rules:** fix import paths of modules in rules package ([fc54a76](https://github.com/salesforce/sa11y/commit/fc54a765953f199ffa1d23580af8244a2ae1d173))
+
+### Features
+
+-   **assert:** add scaffolding for assert package ([0726678](https://github.com/salesforce/sa11y/commit/0726678b78aed455114e00305bbffdbc3ee15739))
+-   **format:** add basic scaffolding for format package ([b28dc2a](https://github.com/salesforce/sa11y/commit/b28dc2a74a512f19537ebcdb630f8ef702ba67f8))
+
 # (2020-03-20)
 
 ### Bug Fixes